fixes async in Customer fetches
