+++ client/services/RestaurantService.js
@@ -13,21 +13,3 @@ export async function getRestaurantById(id) {
     console.error(error);
     }
 }
-
-export async function getFilteredAvailablitiesOfRestaurant(restaurantId, partySize, date, time) {
-  try {
-  const data = await fetch(baseURLAvailabilities + restaurantId + '/filtered?partySize=' + partySize +'&date=' + date + '&time=' + time);
-  return data.json() 
-  } catch (error) {
-    console.error(error);
-  }
-
-}
-export async function getAllAvailablitiesOfRestaurant(restaurantId, date) {
-  try {
-  const data = await fetch(baseURLAvailabilities + restaurantId + '/filtered?date=' + date);
-  return data.json() 
-  } catch (error) {
-    console.error(error);
-  }
-}
